app/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo[srv]==4.7.0
mysql-connector-python==8.4.0
python-dotenv>=0.19.0 
openai>=1.0.0  
pydantic>=1.8.2 